Fix(auth): separated the test file for users and roles
It allows the test to not remove the pre-defined roles in the system.
