--- tests/Feature/UserManagementTest.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Tests\Feature;
-
-use Illuminate\Foundation\Testing\RefreshDatabase;
-use Illuminate\Foundation\Testing\WithFaker;
-use Tests\TestCase;
-
-class UserManagementTest extends TestCase
-{
-
-}
